--- molecule/all/molecule.yml
@@ -1,39 +1,0 @@
----
-dependency:
-  name: galaxy
-driver:
-  name: docker
-lint:
-  name: yamllint
-platforms:
-  - name: instance
-    image: "geerlingguy/docker-${MOLECULE_DISTRO:-ubuntu1804}-ansible:latest"
-    command: ${MOLECULE_DOCKER_COMMAND:-""}
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-    privileged: true
-#    pre_build_image: true
-provisioner:
-  name: ansible
-  lint:
-    name: ansible-lint
-  playbooks:
-    prepare: prepare.yml
-scenario:
-  name: all
-  # test_sequence:
-  #   - lint
-  #   - destroy
-  #   - dependency
-  #   - syntax
-  #   - create
-  #   - prepare
-  #   - converge
-  #   # - idempotence
-  #   - side_effect
-  #   - verify
-  #   - destroy
-verifier:
-  name: testinfra
-  lint:
-    name: flake8
