--- molecule/all/prepare.yml
@@ -1,50 +1,0 @@
----
-- name: Prepare
-  hosts: all
-  pre_tasks:
-    - name: This likely fixes stale apt caches in Docker images
-      apt:
-        update_cache: true
-        cache_valid_time: 3600
-      when:
-        - ansible_os_family == 'Debian'
-
-    - name: Debug
-      debug:
-        msg: "{{ item }}"
-      loop:
-        - "{{ ansible_os_family }}"
-        - "{{ ansible_distribution_major_version }}"
-
-    - name: "[Debian] Configure geerlingguy.mysql with Python2"
-      set_fact:
-        mysql_python_package_debian: python-mysqldb
-      when:
-        - ansible_os_family == 'Debian'
-
-    - name: "[RedHat-7] pre-tasks"
-      block:
-        - name: use MariaDB 10.2
-          yum_repository:
-            name: mariadb
-            description: MariaDB
-            baseurl: http://yum.mariadb.org/10.2/centos7-amd64
-            gpgkey: https://yum.mariadb.org/RPM-GPG-KEY-MariaDB
-            gpgcheck: true
-            state: present
-        - name: "Create pid dir"
-          file:
-            path: "/var/run/mariadb"
-            state: directory
-            owner: root
-            group: root
-            mode: 0757
-            # setype: mysqld_var_run_t
-      when:
-        - ansible_os_family == 'RedHat'
-        - ansible_distribution_major_version == '7'
-
-  roles:
-    - role: geerlingguy.mysql
-      vars:
-        mysql_pid_file: /var/run/mariadb/mariadb.pid
