+++ molecule/default/molecule.yml
@@ -17,8 +17,22 @@ provisioner:
   name: ansible
   lint:
     name: ansible-lint
+  playbooks:
+    prepare: prepare.yml
 scenario:
   name: default
+  # test_sequence:
+  #   - lint
+  #   - destroy
+  #   - dependency
+  #   - syntax
+  #   - create
+  #   - prepare
+  #   - converge
+  #   # - idempotence
+  #   - side_effect
+  #   - verify
+  #   - destroy
 verifier:
   name: testinfra
   lint:
