+++ vars/Debian-18.yml
@@ -1,0 +1,2 @@
+---
+mariadb_backup_package: percona-xtrabackup
